Run 'npm install' with every compilation
We had originally left the 'nodecompile' task as
cljs->js only with the intention that tasks like
'run' and 'install' could manage npm deps on their
own.  However, this proved awkward to actually
develop nodecljs apps since iterative builds
do not use either of those methods.

Therefore, we now run npm-install implicitly with
every compile task.  The biggest downside is that
we will run it even with 'lein install' which is
not strictly necessary.  If this becomes annoying,
we can optimize it later.

Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>